远程ssh功能

+++ src/main/kotlin/julyww/harbor/common/PageResult.kt
@@ -1,6 +1,6 @@
 package julyww.harbor.common
 
 data class PageResult<T>(
-    val total: Int,
+    val total: Long,
     val list: List<T>
 )
